Refactor shared model code into a mixin
Fixes #2268, where the model building code in the model-picker was different from the model page.
