+++ bundles/org.openhab.ui/web/src/pages/settings/model/model.vue
@@ -242,13 +242,10 @@ import ItemDetails from '@/components/model/item-details.vue'
 import MetadataMenu from '@/components/item/metadata/item-metadata-menu.vue'
 import LinkDetails from '@/components/model/link-details.vue'
 
-import { compareItems } from '@/components/widgets/widget-order'
-
-function compareModelItems (o1, o2) {
-  return compareItems(o1.item || o1, o2.item || o2)
-}
+import ModelMixin from '@/pages/settings/model/model-mixin'
 
 export default {
+  mixins: [ModelMixin],
   components: {
     'empty-state-placeholder': () => import('@/components/empty-state-placeholder.vue'),
     ModelDetailsPane,
@@ -260,26 +257,12 @@ export default {
   data () {
     if (!this.$f7.data.model) this.$f7.data.model = {}
     return {
-      ready: false,
-      loading: false,
-      includeNonSemantic: false,
       includeItemName: this.$f7.data.model.includeItemName || false,
       includeItemTags: this.$f7.data.model.includeItemTags || false,
       expanded: this.$f7.data.model.expanded || false,
-      items: [],
-      links: [],
-      locations: [],
-      rootLocations: [],
-      equipment: {},
-      rootEquipment: [],
-      rootPoints: [],
-      rootGroups: [],
-      rootItems: [],
       newItem: null,
       newItemParent: null,
       initSearchbar: false,
-      selectedItem: null,
-      previousSelection: null,
       detailsOpened: false,
       detailsTab: 'state',
       eventSource: null,
@@ -320,7 +303,7 @@ export default {
     modelItem (item) {
       const modelItem = {
         item,
-        opened: null,
+        opened: false,
         class: (item.metadata && item.metadata.semantics) ? item.metadata.semantics.value : '',
         children: {
           locations: [],
@@ -342,47 +325,8 @@ export default {
 
       return modelItem
     },
-    load (update) {
-      // if (this.ready) return
-      this.loading = true
-      const items = this.$oh.api.get('/rest/items?staticDataOnly=true&metadata=.+')
-      const links = this.$oh.api.get('/rest/links')
-      Promise.all([items, links]).then((data) => {
-        this.items = data[0]
-        this.links = data[1]
-
-        if (this.newItem) {
-          this.items.push(this.newItem)
-        }
-
-        this.locations = this.items.filter((i) => i.metadata && i.metadata.semantics && i.metadata.semantics.value.indexOf('Location') === 0)
-        this.equipment = this.items.filter((i) => i.metadata && i.metadata.semantics && i.metadata.semantics.value.indexOf('Equipment') === 0)
-        this.points = this.items.filter((i) => i.metadata && i.metadata.semantics && i.metadata.semantics.value.indexOf('Point') === 0)
-
-        this.rootLocations = this.locations
-          .filter((i) => !i.metadata.semantics.config || !i.metadata.semantics.config.isPartOf)
-          .map(this.modelItem).sort(compareModelItems)
-        this.rootLocations.forEach(this.getChildren)
-        this.rootEquipment = this.equipment
-          .filter((i) => !i.metadata.semantics.config || (!i.metadata.semantics.config.isPartOf && !i.metadata.semantics.config.hasLocation))
-          .map(this.modelItem).sort(compareModelItems)
-        this.rootEquipment.forEach(this.getChildren)
-        this.rootPoints = this.points
-          .filter((i) => !i.metadata.semantics.config || (!i.metadata.semantics.config.isPointOf && !i.metadata.semantics.config.hasLocation))
-          .map(this.modelItem).sort(compareModelItems)
-
-        if (this.includeNonSemantic) {
-          this.rootGroups = this.items
-            .filter((i) => i.type === 'Group' && (!i.metadata || !i.metadata.semantics) && i.groupNames.length === 0)
-            .map(this.modelItem).sort(compareModelItems)
-          this.rootGroups.forEach(this.getChildren)
-          this.rootItems = this.items
-            .filter((i) => i.type !== 'Group' && (!i.metadata || !i.metadata.semantics) && i.groupNames.length === 0)
-            .map(this.modelItem).sort(compareModelItems)
-        }
-
-        this.loading = false
-        this.ready = true
+    load () {
+      this.loadModel().then(() => {
         this.$nextTick(() => {
           this.initSearchbar = true
           this.applyExpandedOption()
@@ -412,59 +356,6 @@ export default {
     stopEventSource () {
       this.$oh.sse.close(this.eventSource)
       this.eventSource = null
-    },
-    getChildren (parent) {
-      // open the parent node of the placeholder
-      if (this.newItemParent && this.newItemParent === parent.item.name) {
-        parent.opened = true
-      }
-
-      // restore previous selection
-      if (this.previousSelection && parent.item.name === this.previousSelection.item.name) {
-        this.selectedItem = parent
-        this.previousSelection = null
-        this.selectItem(parent)
-      }
-
-      if (parent.class.indexOf('Location') === 0) {
-        parent.children.locations = this.locations
-          .filter((i) => i.metadata.semantics.config && i.metadata.semantics.config.isPartOf === parent.item.name)
-          .map(this.modelItem).sort(compareModelItems)
-        parent.children.locations.forEach(this.getChildren)
-        parent.children.equipment = this.equipment
-          .filter((i) => i.metadata.semantics.config && i.metadata.semantics.config.hasLocation === parent.item.name)
-          .map(this.modelItem).sort(compareModelItems)
-        parent.children.equipment.forEach(this.getChildren)
-
-        parent.children.points = this.points
-          .filter((i) => i.metadata.semantics.config && i.metadata.semantics.config.hasLocation === parent.item.name)
-          .map(this.modelItem).sort(compareModelItems)
-      } else {
-        parent.children.equipment = this.equipment
-          .filter((i) => i.metadata.semantics.config && i.metadata.semantics.config.isPartOf === parent.item.name)
-          .map(this.modelItem).sort(compareModelItems)
-        parent.children.equipment.forEach(this.getChildren)
-
-        parent.children.points = this.points
-          .filter((i) => i.metadata.semantics.config && i.metadata.semantics.config.isPointOf === parent.item.name)
-          .map(this.modelItem).sort(compareModelItems)
-      }
-
-      if (this.includeNonSemantic) {
-        parent.children.groups = this.items
-          .filter((i) => i.type === 'Group' && (!i.metadata || (i.metadata && !i.metadata.semantics)) && i.groupNames.indexOf(parent.item.name) >= 0)
-          .map(this.modelItem).sort(compareModelItems)
-        parent.children.groups.forEach(this.getChildren)
-        if (parent.item.metadata && parent.item.metadata.semantics) {
-          parent.children.items = this.items
-            .filter((i) => i.type !== 'Group' && (!i.metadata || (i.metadata && !i.metadata.semantics)) && i.groupNames.indexOf(parent.item.name) >= 0)
-            .map(this.modelItem).sort(compareModelItems)
-        } else {
-          parent.children.items = this.items
-            .filter((i) => i.type !== 'Group' && i.groupNames.indexOf(parent.item.name) >= 0)
-            .map(this.modelItem).sort(compareModelItems)
-        }
-      }
     },
     selectItem (item) {
       this.selectedItem = item
@@ -505,19 +396,6 @@ export default {
       this.expanded = !this.expanded
       this.$f7.data.model.expanded = this.expanded
       this.applyExpandedOption()
-    },
-    applyExpandedOption () {
-      const treeviewItems = document.querySelectorAll('.treeview-item')
-
-      treeviewItems.forEach(item => {
-        if (item.classList.contains('treeview-item')) {
-          if (this.expanded) {
-            item.classList.add('treeview-item-opened')
-          } else {
-            item.classList.remove('treeview-item-opened')
-          }
-        }
-      })
     },
     addSemanticItem (semanticType) {
       this.newItem = {
